2023/SOSSF/requirements.txt
absl-py==0.10.0
affine==2.3.0
albumentations==1.0.0
asn1crypto==0.24.0
astor==0.7.1
astunparse==1.6.3
attrs==18.2.0
backcall==0.1.0
backports.weakref==1.0rc1
bleach==1.5.0
blessings==1.7
blis==0.4.1
cached-property==1.5.2
cachetools==4.1.1
catalogue==1.0.0
certifi==2021.5.30
cffi==1.12.3
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
cloudpickle==0.7.0
connectivity==31
cryptography==2.5
cycler==0.10.0
cymem==2.0.3
Cython==0.29.17
cytoolz==0.9.0.1
dask==1.1.1
decorator==4.3.2
defusedxml==0.5.0
easydict==1.9
entrypoints==0.3
flatbuffers==1.12
future==0.18.2
gast==0.2.2
GDAL==3.0.2
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gpustat==0.6.0
graphviz==0.19.1
grpcio==1.34.1
h5py==2.10.0
html5lib==0.9999999
icc-rt==2020.0.133
idna==2.8
imageio==2.3.0
imantics==0.1.12
imgaug==0.4.0
importlib-metadata==2.0.0
intel-openmp==2020.0.133
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.2
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.3.1
Keras-Applications==1.0.8
keras-contrib @ git+https://www.github.com/keras-team/keras-contrib.git@3fc5ef709e061416f4bc8a92ca3750c824b5d2b0
keras-layer-normalization==0.16.0
keras-models==0.0.7
Keras-Preprocessing==1.1.2
kiwisolver==1.0.1
libtiff==0.4.2
lightgbm==3.2.1
lxml==4.9.2
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
mkl==2019.0
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==3.0.5
murmurhash==1.0.2
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
notebook==5.7.5
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603487797006/work
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.0.0.21
opencv-python==4.0.0.21
opencv-python-headless==4.5.2.54
opt-einsum==3.3.0
pandas==0.24.1
pandocfilters==1.4.2
parso==0.3.2
pathlib==1.0.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
plac==1.1.3
preshed==3.0.2
prometheus-client==0.5.0
prompt-toolkit==2.0.8
protobuf==3.12.2
psutil==5.7.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycococreatortools==0.2.0
pycocotools==2.0.6
pycosat==0.6.3
pycparser==2.19
pydensecrf==1.0rc3
pydot==1.4.2
pydot-ng==2.0.0
pydotplus==2.0.2
Pygments==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.3.1
pyrsistent==0.14.11
pyshp==2.1.3
PySocks==1.6.8
python-dateutil==2.7.5
pytz==2018.9
PyWavelets==1.1.1
PyYAML==5.1
pyzmq==17.1.2
qtconsole==4.4.3
rasterio==1.1.2
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.15.71
scikit-image==0.17.2
scikit-learn==0.20.3
scipy==1.4.1
Send2Trash==1.5.0
Shapely==1.7.1
simplegeneric==0.8.1
six @ file:///tmp/build/80754af9/six_1623709665295/work
sklearn==0.0
snuggs==1.4.7
spacy==2.3.2
srsly==1.0.2
tbb==2019.0
tbb4py==2019.0
tensorboard==1.15.0
tensorboardX==2.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
Theano==1.0.4
thinc==7.4.1
tifffile==2020.9.3
toolz==0.9.0
toposort==1.5
torch==1.1.0
torchvision==0.3.0
tornado==5.1.1
tqdm==4.50.2
traitlets==4.3.2
typing==3.6.4
typing-extensions==3.7.4.3
unrar==0.4
urllib3==1.24.1
wasabi==0.8.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.12.1
xmljson==0.2.1
zipp==3.3.0
